requirements.txt
numpy>=1.26
pandas>=2.0
pyyaml>=6.0
matplotlib>=3.8
